Added parse.sh
parse.sh is a shell script for creating the input file in necessary format
for the programs included in EmoDetect. This shell script recursively goes
through all subdirectories, and lists all .jpg files along with a label to
indicate the associated emotion. This file has been optimised to work with
the MUG facial expressions data set mentioned in the README. This file may
need to be changed if another data set is used.